Integrated sigin logic at login full screen
