+++ login-workflow/package.json
@@ -62,6 +62,7 @@
     "peerDependencies": {
         "@brightlayer-ui/react-native-components": "^8.0.4",
         "@brightlayer-ui/react-native-vector-icons": "^2.0.0",
+        "@okta/okta-react-native": "^2.12.0",
         "i18next": "^23.11.5",
         "react": "^18.0.0",
         "react-i18next": "^14.1.2",
